Category id type 변경
Long -> Integer

+++ src/main/java/com/widehouse/cafe/domain/cafe/CafeRepository.java
@@ -4,7 +4,6 @@ import java.util.List;
 
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
 /**
@@ -18,8 +17,8 @@ public interface CafeRepository extends CrudRepository<Cafe, Long> {
     @Cacheable("cafes_category")
     List<Cafe> findByCategory(Category category, Pageable pageable);
 
-    List<Cafe> findByCategoryId(Long categoryId);
+    List<Cafe> findByCategoryId(Integer categoryId);
 
     @Cacheable("cafes_category")
-    List<Cafe> findByCategoryId(Long categoryId, Pageable pageable);
+    List<Cafe> findByCategoryId(Integer categoryId, Pageable pageable);
 }
